Readme done and styled renaming
