--- CHANGELOG.md
@@ -1,17 +1,0 @@
-### 3 July 7, 2020
-
-- Updated to React Router v5.2.0 with new Syntax and Hooks: No more usaged of the `component` property on the Route, new hooks useHistory and useLocation.
-
-### Oct 23, 2020
-
-- Added support for async/wait using the `@babel/plugin-transform-runtime` plugin. No more regenerator/runtime issue.
-- Replaced now.json with vercel.json
-
-### December 9, 2021
-
-- Support for node 16 and droppped 14.
-- Updated all libraries to the latest
-- Now the prettier us being used as a vscode plugin instead.
-- Dropped eslint (for now)
-
-
