+++ src/js/components/FiltersColumn/StyledFilterComponents/FilterLabelStyled.js
